cli: uint is not always defined, use long int for the response timeout instead
gh-3

+++ src/sf_serialmaccli.h
@@ -76,7 +76,7 @@ namespace sf {
             std::map<std::string, docopt::value> args;
             SerialMACConfig *serialMACConfig = nullptr;
             SerialPortConfig *serialPortConfig = nullptr;
-            uint respTimeoutSecs = 5;
+            long int respTimeoutSecs = 5;
 
             bool interactive;
             bool noInvertedLengthField;
